daf-processor/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
json-repair>=0.30.0
